#36 Renamed settings package to persistence
