+++ src/main/kotlin/com/github/stephenhamiltonc/timecard_web/core/persistence/EntryLifespan.kt
@@ -1,4 +1,4 @@
-package com.github.stephenhamiltonc.timecard_web.core.settings
+package com.github.stephenhamiltonc.timecard_web.core.persistence
 
 import io.kvision.core.StringPair
 import kotlinx.serialization.Serializable
