+++ src/main/kotlin/com/github/stephenhamiltonc/timecard_web/page/home/Controls.kt
@@ -4,8 +4,8 @@ import com.github.stephenhamiltonc.timecard.Timecard
 import com.github.stephenhamiltonc.timecard_web.card
 import com.github.stephenhamiltonc.timecard_web.core.format
 import com.github.stephenhamiltonc.timecard_web.core.formatMinutes
-import com.github.stephenhamiltonc.timecard_web.core.settings.Persistence
-import com.github.stephenhamiltonc.timecard_web.core.settings.TimeFormat
+import com.github.stephenhamiltonc.timecard_web.core.persistence.Persistence
+import com.github.stephenhamiltonc.timecard_web.core.persistence.TimeFormat
 import io.kvision.core.Container
 import io.kvision.core.TooltipOptions
 import io.kvision.core.enableTooltip
